Expanding tile for numbers > 10

+++ Assets/Scripts/DigitValue.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public class DigitValue : MonoBehaviour
 {
     public int integerValue;
-    public int orderOfMagnitude;
+    public double orderOfMagnitude;
 }
